+++ spec/lib/udongo/meta_info_spec.rb
@@ -1,0 +1,12 @@
+require 'rails_helper'
+
+describe Udongo::MetaInfo do
+  let(:instance) { Udongo::MetaInfo.new }
+
+  it '#respond_to?' do
+    expect(instance).to respond_to(
+      :keywords, :keywords=, :description, :description=, :custom, :custom=,
+      :title, :title=
+    )
+  end
+end